Rename jlink/swd.rs to arm.rs
